requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
python-dotenv==0.19.2
